git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/gsoc_bzauthd@19840 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/NetHandler.h
@@ -1,9 +1,9 @@
 /* bzflag
- * Copyright (c) 1993 - 2008 Tim Riker
+ * Copyright (c) 1993 - 2009 Tim Riker
  *
  * This package is free software;  you can redistribute it and/or
  * modify it under the terms of the license found in the file
- * named LICENSE that should have accompanied this file.
+ * named COPYING that should have accompanied this file.
  *
  * THIS PACKAGE IS PROVIDED ``AS IS'' AND WITHOUT ANY EXPRESS OR
  * IMPLIED WARRANTIES, INCLUDING, WITHOUT LIMITATION, THE IMPLIED
@@ -120,7 +120,7 @@ public:
   */
   RxStatus    tcpReceive(bool doCodes = true);
   void       *getTcpBuffer();
-  unsigned int getTcpReadSize ( void  );
+  size_t      getTcpReadSize ( void  );
 
   /// Request if there is any buffered udp messages waiting to be sent
   static bool	anyUDPPending() {return pendingUDP;};
@@ -203,7 +203,7 @@ private:
   /// current TCP msg
   char tcpmsg[MaxPacketLen];
   /// bytes read in current msg
-  int tcplen;
+  size_t tcplen;
   /// current UDP msg
   static char	       udpmsg[MaxPacketLen];
   static int		udpLen;
